backend/pyproject.toml
[tool.poetry]
name = "holly-backend"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.10"
strawberry-graphql = {extras = ["debug-server"], version = "^0.199.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
